+++ src/main/java/org/morphix/reflection/Methods.java
@@ -12,8 +12,6 @@
  */
 package org.morphix.reflection;
 
-import static org.morphix.reflection.predicates.MemberPredicates.withAnnotation;
-
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -31,7 +29,7 @@ import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 import org.morphix.lang.JavaObjects;
-import org.morphix.lang.Messages;
+import org.morphix.reflection.predicates.MemberPredicates;
 
 /**
  * Utility reflection static methods for java methods. The philosophy of this class is to provide methods that cover
@@ -177,16 +175,15 @@ public interface Methods {
 	static <T extends Type> T getGenericReturnType(final Method method, final int index) {
 		Type type = method.getGenericReturnType();
 		if (!(type instanceof ParameterizedType parameterizedType)) {
-			throw new ReflectionException(Messages.message(
+			throw new ReflectionException(
 					"{} is a raw return type for method {}.{}",
-					type.getTypeName(), method.getDeclaringClass().getCanonicalName(), method.getName()));
+					type.getTypeName(), method.getDeclaringClass().getCanonicalName(), method.getName());
 		}
 		Type[] actualTypeArguments = parameterizedType.getActualTypeArguments();
 		if (index >= actualTypeArguments.length) {
-			throw new ReflectionException(Messages.message(
+			throw new ReflectionException(
 					"Could not find generic argument at index {} for generic return type {} with {} generic argument(s) for method {}.{}",
-					index, parameterizedType.getTypeName(), actualTypeArguments.length, method.getDeclaringClass().getCanonicalName(),
-					method.getName()));
+					index, parameterizedType.getTypeName(), actualTypeArguments.length, method.getDeclaringClass().getCanonicalName(), method.getName());
 		}
 		Type returnType = actualTypeArguments[index];
 		return JavaObjects.cast(returnType);
@@ -216,9 +213,9 @@ public interface Methods {
 		try {
 			return getGenericReturnType(method, index);
 		} catch (ClassCastException e) {
-			throw new ReflectionException(Messages.message(
+			throw new ReflectionException(
 					"Could not infer actual generic return type argument from {} for method {}.{}",
-					method.getGenericReturnType().getTypeName(), method.getDeclaringClass().getCanonicalName(), method.getName()), e);
+					method.getGenericReturnType().getTypeName(), method.getDeclaringClass().getCanonicalName(), method.getName());
 		}
 	}
 
@@ -352,15 +349,13 @@ public interface Methods {
 		try {
 			primitiveFieldType = Primitives.toPrimitive(field.getType());
 		} catch (ReflectionException e) {
-			throw new ReflectionException(Messages.message(
-					"Error finding method: {}({})",
-					methodName, field.getType().getCanonicalName()), e);
+			throw new ReflectionException(ErrorMessage.ERROR_FINDING_METHOD, methodName, field.getType().getCanonicalName());
 		}
 		method = Methods.getOneDeclaredInHierarchy(methodName, cls, primitiveFieldType);
 		if (null == method) {
-			throw new ReflectionException(Messages.message(
+			throw new ReflectionException(
 					"Error finding method: {}({}) or {}({})",
-					methodName, field.getType().getCanonicalName(), methodName, primitiveFieldType.getCanonicalName()));
+					methodName, field.getType().getCanonicalName(), methodName, primitiveFieldType.getCanonicalName());
 		}
 		return method;
 	}
@@ -379,15 +374,13 @@ public interface Methods {
 		for (Method method : cls.getMethods()) {
 			if (Modifier.isAbstract(method.getModifiers())) {
 				if (null != singleAbstractMethod) {
-					throw new ReflectionException(Messages.message(
-							"{} is not a functional interface because it has more than one abstract method", cls));
+					throw new ReflectionException("{} is not a functional interface because it has more than one abstract method", cls);
 				}
 				singleAbstractMethod = method;
 			}
 		}
 		if (null == singleAbstractMethod) {
-			throw new ReflectionException(Messages.message(
-					"{} is not a functional interface because it has no abstract method", cls));
+			throw new ReflectionException("{} is not a functional interface because it has no abstract method", cls);
 		}
 		return singleAbstractMethod;
 	}
@@ -413,16 +406,39 @@ public interface Methods {
 			if (null != obj) {
 				className = obj instanceof Class<?> cls ? cls.getCanonicalName() : obj.getClass().getCanonicalName();
 			}
-			throw new ReflectionException(Messages.message(
-					"Error invoking {}.{}: {}.", className, method.getName(), cause.getMessage()), e);
+			throw new ReflectionException(e, ErrorMessage.ERROR_INVOKING_METHOD, className, method.getName(), cause.getMessage());
 		} catch (Exception e) {
 			// escalate any exception invoking the method
 			String className = method.getDeclaringClass().getCanonicalName();
 			if (null != obj) {
 				className = obj instanceof Class<?> cls ? cls.getCanonicalName() : obj.getClass().getCanonicalName();
 			}
-			throw new ReflectionException(Messages.message(
-					"Error invoking {}.{}: {}.", className, method.getName(), e.getMessage()), e);
+			throw new ReflectionException(e, ErrorMessage.ERROR_INVOKING_METHOD, className, method.getName(), e.getMessage());
+		}
+	}
+
+	/**
+	 * Namespace for error messages. These messages are used for exception messages.
+	 *
+	 * @author Radu Sebastian LAZIN
+	 */
+	static class ErrorMessage {
+
+		/**
+		 * Error invoking method message.
+		 */
+		public static final String ERROR_INVOKING_METHOD = "Error invoking method {}.{}: {}.";
+
+		/**
+		 * Error finding method message.
+		 */
+		public static final String ERROR_FINDING_METHOD = "Error finding method: {}({})";
+
+		/**
+		 * Private constructor to avoid instantiation.
+		 */
+		private ErrorMessage() {
+			throw Constructors.unsupportedOperationException();
 		}
 	}
 
@@ -455,16 +471,14 @@ public interface Methods {
 				if (null != obj) {
 					className = obj instanceof Class<?> cls ? cls.getCanonicalName() : obj.getClass().getCanonicalName();
 				}
-				throw new ReflectionException(Messages.message(
-						"Error invoking {}.{}: {}.", className, method.getName(), cause.getMessage()), e);
+				throw new ReflectionException(e, ErrorMessage.ERROR_INVOKING_METHOD, className, method.getName(), cause.getMessage());
 			} catch (Exception e) {
 				// escalate any exception invoking the method
 				String className = method.getDeclaringClass().getCanonicalName();
 				if (null != obj) {
 					className = obj instanceof Class<?> cls ? cls.getCanonicalName() : obj.getClass().getCanonicalName();
 				}
-				throw new ReflectionException(Messages.message(
-						"Error invoking {}.{}: {}.", className, method.getName(), e.getMessage()), e);
+				throw new ReflectionException(e, ErrorMessage.ERROR_INVOKING_METHOD, className, method.getName(), e.getMessage());
 			}
 		}
 
@@ -480,7 +494,7 @@ public interface Methods {
 		 * @throws ReflectionException if any error occurs during method invocation
 		 */
 		static <T, A extends Annotation> void invokeWithAnnotation(final T obj, final Class<A> annotationClass) {
-			List<Method> methods = Methods.getAllDeclaredInHierarchy(obj.getClass(), withAnnotation(annotationClass));
+			List<Method> methods = Methods.getAllDeclaredInHierarchy(obj.getClass(), MemberPredicates.withAnnotation(annotationClass));
 			for (Method method : methods) {
 				IgnoreAccess.invoke(method, obj);
 			}